+++ src/app/popup/popup.module.scss
@@ -1,0 +1,19 @@
+.container {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: #00000080;
+  .popup {
+    background-color: white;
+    padding: 1em 2em 2em 2em;
+    border-radius: 10px;
+    box-shadow: 2px 2px 8px #00000080;
+    max-width: 100%;
+    overflow-x: scroll;
+  }
+}
